+++ source/ChanSort.Loader.PhilipsXml/Channel.cs
@@ -8,6 +8,7 @@ namespace ChanSort.Loader.PhilipsXml
     public readonly XmlNode SetupNode;
     public string RawName;
     public string RawSatellite;
+    public int Format;
 
     internal Channel(SignalSource source, int order, int rowId, XmlNode setupNode)
     {