+++ source/ChanSort/GlobalImageCollection.cs
@@ -97,6 +97,7 @@ namespace ChanSort.Ui
       this.sharedImageCollection1.ImageSource.Images.SetKeyName(41, "0041.png");
       this.sharedImageCollection1.ImageSource.Images.SetKeyName(42, "0042.png");
       this.sharedImageCollection1.ImageSource.Images.SetKeyName(43, "0043.png");
+      this.sharedImageCollection1.ImageSource.Images.SetKeyName(44, "0044.png");
       ((System.ComponentModel.ISupportInitialize)(this.sharedImageCollection1.ImageSource)).EndInit();
       ((System.ComponentModel.ISupportInitialize)(this.sharedImageCollection1)).EndInit();