+++ source/ChanSort/Printing/ChannelListReport.cs
@@ -12,16 +12,24 @@ namespace ChanSort.Ui
     private readonly int subListIndex;
     private readonly bool orderByName;
 
-    public ChannelListReport(ChannelList list, int subListIndex, bool orderByName, Font font)
+    public ChannelListReport(ChannelList list, int subListIndex, bool orderByName, Font font, int columns)
     {
       this.subListIndex = subListIndex;
       this.orderByName = orderByName;
+      if (columns < 1) 
+        columns = 1;
 
       InitializeComponent();
       this.DataSource = GenerateDataSource(list);
 
       this.txtHeading.Text = list.ShortCaption + (subListIndex <= 0 ? "" : " - Fav " + (char)('A' + subListIndex - 1));
-     
+      this.txtNumber.WidthF = this.txtNumber.Width * font.SizeInPoints / 8;
+      this.bandChannelDetail.MultiColumn.ColumnCount = columns;
+      this.bandChannelDetail.MultiColumn.ColumnWidth =  (this.PageSize.Width - this.Margins.Left - this.Margins.Right - (columns - 1) * this.bandChannelDetail.MultiColumn.ColumnSpacing) / columns;
+
+      this.txtChannelName.LeftF = this.txtNumber.RightF;
+      this.txtChannelName.Width = (int)(this.bandChannelDetail.MultiColumn.ColumnWidth - this.txtChannelName.LeftF);
+
       this.txtHeading.Font = new Font(font.Name, font.Size+4, FontStyle.Bold);
       this.txtNumber.Font = font;
       this.txtChannelName.Font = font;